fixed auto-sign when reply to signed
