+++ extension/css/cryptup.css
@@ -1667,6 +1667,7 @@ table#compose td.bottom div.sending-container div.sending-options-container {
   z-index: 1;
   box-shadow: 0 0 4px 0 rgba(0, 0, 0, 0.2);
   min-width: 256px;
+  bottom: 34px;
 }
 
 table#compose td.bottom div.sending-container.popover-opened div.sending-options-container {
